# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LoginDto": {
				"properties": {
					"Email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"Email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRole": {
				"enum": [
					"ADMIN",
					"PASSENGERS",
					"DRIVER"
				],
				"type": "string"
			},
			"SignupDto": {
				"properties": {
					"FullName": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"FullName",
					"Email",
					"password",
					"Phone",
					"Role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Userstatus": {
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"BLOCKED"
				],
				"type": "string"
			},
			"Position": {
				"items": {
					"type": "number",
					"format": "double"
				},
				"type": "array",
				"description": "Position object.\nhttps://datatracker.ietf.org/doc/html/rfc7946#section-3.1.1"
			},
			"Point": {
				"properties": {
					"coordinates": {
						"$ref": "#/components/schemas/Position"
					},
					"type": {
						"type": "string",
						"enum": [
							"Point"
						],
						"nullable": false
					}
				},
				"required": [
					"coordinates",
					"type"
				],
				"type": "object",
				"description": "Point geometry object.\nhttps://datatracker.ietf.org/doc/html/rfc7946#section-3.1.2"
			},
			"LineString": {
				"properties": {
					"coordinates": {
						"items": {
							"$ref": "#/components/schemas/Position"
						},
						"type": "array"
					},
					"type": {
						"type": "string",
						"enum": [
							"LineString"
						],
						"nullable": false
					}
				},
				"required": [
					"coordinates",
					"type"
				],
				"type": "object",
				"description": "LineString geometry object.\nhttps://datatracker.ietf.org/doc/html/rfc7946#section-3.1.4"
			},
			"VehicleType": {
				"enum": [
					"BIKE",
					"CAR",
					"ELECTRIC",
					"AUTO"
				],
				"type": "string"
			},
			"RideStatus": {
				"enum": [
					"REQUESTED",
					"SEARCHING",
					"ACCEPTED",
					"ARRIVING",
					"STARTED",
					"COMPLETED",
					"CANCELED",
					"DRIVERNOTFOUND"
				],
				"type": "string"
			},
			"Cancledby": {
				"enum": [
					"DRIVER",
					"PASSENGERS",
					"SYSTEM"
				],
				"type": "string"
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"DeletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"FullName": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Userstatus"
					},
					"address": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"Role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"rides": {
						"items": {
							"$ref": "#/components/schemas/Ride"
						},
						"type": "array"
					},
					"Driver": {
						"$ref": "#/components/schemas/Driver"
					}
				},
				"required": [
					"id",
					"DeletedAt",
					"UpdatedAt",
					"CreatedAt",
					"FullName",
					"Email",
					"password",
					"Phone",
					"profile",
					"status",
					"address",
					"Role",
					"rides",
					"Driver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Driverstatus": {
				"enum": [
					"ONLINE",
					"OFFLINE",
					"BUSY"
				],
				"type": "string"
			},
			"Ride": {
				"properties": {
					"id": {
						"type": "string"
					},
					"DeletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"pickupAddress": {
						"type": "string"
					},
					"pickupLocation": {
						"$ref": "#/components/schemas/Point"
					},
					"DropoffAddress": {
						"type": "string"
					},
					"DropoffLocation": {
						"$ref": "#/components/schemas/Point"
					},
					"routepath": {
						"$ref": "#/components/schemas/LineString"
					},
					"estimatedFare": {
						"type": "number",
						"format": "double"
					},
					"FinalFare": {
						"type": "number",
						"format": "double"
					},
					"estimatedDistance": {
						"type": "number",
						"format": "double"
					},
					"ActualDistance": {
						"type": "number",
						"format": "double"
					},
					"DurationMinutes": {
						"type": "number",
						"format": "double"
					},
					"reqVehicleType": {
						"$ref": "#/components/schemas/VehicleType"
					},
					"ridestauts": {
						"$ref": "#/components/schemas/RideStatus"
					},
					"cancledby": {
						"$ref": "#/components/schemas/Cancledby"
					},
					"cancellationResion": {
						"type": "string"
					},
					"cancelAt": {
						"type": "string",
						"format": "date-time"
					},
					"driverAcceptedAt": {
						"type": "string",
						"format": "date-time"
					},
					"driverArrivedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tripStartAt": {
						"type": "string",
						"format": "date-time"
					},
					"tripEndAt": {
						"type": "string",
						"format": "date-time"
					},
					"promoCode": {
						"type": "string"
					},
					"isScheduled": {
						"type": "boolean"
					},
					"ScheduledAt": {
						"type": "string",
						"format": "date-time"
					},
					"specialInstruction": {
						"type": "string"
					},
					"rider": {
						"$ref": "#/components/schemas/User"
					},
					"vechicle": {
						"$ref": "#/components/schemas/Vechicles"
					},
					"driver": {
						"$ref": "#/components/schemas/Driver"
					}
				},
				"required": [
					"id",
					"DeletedAt",
					"UpdatedAt",
					"CreatedAt",
					"pickupAddress",
					"pickupLocation",
					"DropoffAddress",
					"DropoffLocation",
					"routepath",
					"estimatedFare",
					"FinalFare",
					"estimatedDistance",
					"ActualDistance",
					"DurationMinutes",
					"reqVehicleType",
					"ridestauts",
					"cancledby",
					"cancellationResion",
					"cancelAt",
					"driverAcceptedAt",
					"driverArrivedAt",
					"tripStartAt",
					"tripEndAt",
					"rider",
					"vechicle",
					"driver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Vechicles": {
				"properties": {
					"id": {
						"type": "string"
					},
					"DeletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"model": {
						"type": "string"
					},
					"plateNumber": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					},
					"type": {
						"$ref": "#/components/schemas/VehicleType"
					},
					"seatCapacity": {
						"type": "string"
					},
					"driver": {
						"$ref": "#/components/schemas/Driver"
					},
					"rides": {
						"items": {
							"$ref": "#/components/schemas/Ride"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"DeletedAt",
					"UpdatedAt",
					"CreatedAt",
					"model",
					"plateNumber",
					"isDefault",
					"type",
					"seatCapacity",
					"driver",
					"rides"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Driver": {
				"properties": {
					"id": {
						"type": "string"
					},
					"DeletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"licenseNumber": {
						"type": "string"
					},
					"licenseExpery": {
						"type": "string",
						"format": "date-time"
					},
					"citizenshipNumber": {
						"type": "string"
					},
					"currentBearing": {
						"type": "number",
						"format": "double"
					},
					"lastLocationUpdate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/Driverstatus"
					},
					"totaltrip": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"isApproped": {
						"type": "boolean"
					},
					"CurrentLocation": {
						"$ref": "#/components/schemas/Point"
					},
					"ridesasDriver": {
						"items": {
							"$ref": "#/components/schemas/Ride"
						},
						"type": "array"
					},
					"vechicles": {
						"$ref": "#/components/schemas/Vechicles"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"DeletedAt",
					"UpdatedAt",
					"CreatedAt",
					"licenseNumber",
					"licenseExpery",
					"citizenshipNumber",
					"currentBearing",
					"lastLocationUpdate",
					"status",
					"totaltrip",
					"rating",
					"isApproped",
					"CurrentLocation",
					"ridesasDriver",
					"vechicles",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRideDto": {
				"properties": {
					"PickupAddress": {
						"type": "string"
					},
					"pickuplat": {
						"type": "number",
						"format": "double"
					},
					"pickuplng": {
						"type": "number",
						"format": "double"
					},
					"DropoffAddress": {
						"type": "string"
					},
					"dropofflat": {
						"type": "number",
						"format": "double"
					},
					"dropofflng": {
						"type": "number",
						"format": "double"
					},
					"vehicleType": {
						"$ref": "#/components/schemas/VehicleType"
					},
					"PromoCode": {
						"type": "string"
					},
					"SpecialInstruction": {
						"type": "string"
					},
					"isScheduled": {
						"type": "boolean"
					},
					"ScheduledAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"PickupAddress",
					"pickuplat",
					"pickuplng",
					"DropoffAddress",
					"dropofflat",
					"dropofflng",
					"vehicleType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Driverdto": {
				"properties": {
					"licenseNumber": {
						"type": "string"
					},
					"licenseExpery": {
						"type": "string",
						"format": "date-time"
					},
					"citizenshipNumber": {
						"type": "string"
					},
					"vehicleType": {
						"$ref": "#/components/schemas/VehicleType"
					},
					"Vechiclemodel": {
						"type": "string"
					},
					"plateNumber": {
						"type": "string"
					},
					"seatCapacity": {
						"type": "string"
					}
				},
				"required": [
					"licenseNumber",
					"licenseExpery",
					"citizenshipNumber",
					"vehicleType",
					"Vechiclemodel",
					"plateNumber",
					"seatCapacity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				}
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupDto"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/auth/checkauth": {
			"get": {
				"operationId": "Checkauth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/getusers": {
			"get": {
				"operationId": "Getuser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalpages": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/User"
											},
											"type": "array"
										}
									},
									"required": [
										"totalpages",
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "Role",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UserRole"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/deleteuser/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ride/estimatefare": {
			"get": {
				"operationId": "Estimatefare",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"estimatedDurationMinutes": {
											"type": "number",
											"format": "double"
										},
										"estimatedDistanceKm": {
											"type": "number",
											"format": "double"
										},
										"estimatedFare": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"estimatedDurationMinutes",
										"estimatedDistanceKm",
										"estimatedFare"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pickuplat",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pickuplng",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dropofflat",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dropofflng",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "vehicleType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/VehicleType"
						}
					}
				]
			}
		},
		"/ride/createride": {
			"post": {
				"operationId": "CreateRide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ride"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRideDto"
							}
						}
					}
				}
			}
		},
		"/ride/availableRides": {
			"get": {
				"operationId": "GetAvailableRides",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Ride"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ride/active": {
			"get": {
				"operationId": "GetActiveRide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ride"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/ride/{id}/accept": {
			"patch": {
				"operationId": "AcceptRides",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ride/getacceptRide": {
			"get": {
				"operationId": "GetacceptRide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ride"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "rideId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ride/complete/{rideId}": {
			"patch": {
				"operationId": "CompleteRide",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "rideId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ride/getstatus": {
			"get": {
				"operationId": "Status",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RideStatus"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "rideId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ride/getuserride": {
			"get": {
				"operationId": "GetuserRide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Ride"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ride/GetDriver": {
			"get": {
				"operationId": "GetDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ride"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "RideId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/geocoading/geocode": {
			"get": {
				"operationId": "GeocoadeAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/geocoading/reversecode": {
			"get": {
				"operationId": "ReversegeoAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lat",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "lng",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/driver/registerdriver": {
			"post": {
				"operationId": "RegisterDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"vehicle": {
											"properties": {
												"seatCapacity": {
													"type": "string"
												},
												"type": {
													"$ref": "#/components/schemas/VehicleType"
												},
												"plateNumber": {
													"type": "string"
												},
												"model": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"seatCapacity",
												"type",
												"plateNumber",
												"model",
												"id"
											],
											"type": "object"
										},
										"driver": {
											"properties": {
												"citizenshipNumber": {
													"type": "string"
												},
												"licenseNumber": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"citizenshipNumber",
												"licenseNumber",
												"id"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"vehicle",
										"driver",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Driverdto"
							}
						}
					}
				}
			}
		},
		"/driver/myprofile": {
			"get": {
				"operationId": "GetDriverProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Driver"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/driver/location": {
			"post": {
				"operationId": "DriverLocation",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"bearing": {
										"type": "number",
										"format": "double"
									},
									"lng": {
										"type": "number",
										"format": "double"
									},
									"lat": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"bearing",
									"lng",
									"lat"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/driver/getapprovedDriver": {
			"get": {
				"operationId": "GetApprovedDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Driver"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/driver/approved/{id}": {
			"patch": {
				"operationId": "ApprovedDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}